+++ src/models/ftl-animation-sheet.ts
@@ -1,0 +1,5 @@
+import {FtlValue} from './ftl-value';
+
+export class FtlAnimationSheet extends FtlValue {
+  readonly kind = 'animation sheet';
+}