+++ src/models/ftl-animation.ts
@@ -1,0 +1,5 @@
+import {FtlValue} from './ftl-value';
+
+export class FtlAnimation extends FtlValue {
+  kind = 'animation';
+}