+++ src/models/ftl-weapon-animation.ts
@@ -1,0 +1,5 @@
+import {FtlValue} from './ftl-value';
+
+export class FtlWeaponAnimation extends FtlValue {
+  readonly kind = 'weapon animation';
+}