+++ src/providers/ftl-data-provider.ts
@@ -11,6 +11,7 @@ import {RefMapperBase} from '../ref-mappers/ref-mapper';
 import {FtlRoot} from '../models/ftl-root';
 import {ImgPathNames, MusicPaths, SoundWavePaths} from '../data/autocomplete-value-sets';
 import {defaultImgFiles} from '../data/default-img-files';
+import {defaultSoundFiles} from '../data/default-sound-files';
 
 export class FtlDataProvider implements IHTMLDataProvider {
   constructor(onParsed: Event<FtlRoot>, private mappers: RefMapperBase[]) {
@@ -23,7 +24,11 @@ export class FtlDataProvider implements IHTMLDataProvider {
 
   updateFtlData(root: FtlRoot) {
     SoundWavePaths.values.length = 0;
-    SoundWavePaths.values.push(...root.soundWaveFiles.map((file) => ({name: file.modPath})));
+    SoundWavePaths.values.push(...root.soundWaveFiles
+        .map((file) => file.modPath)
+        .concat(defaultSoundFiles)
+        .map((modPath) => ({name: modPath})));
+
     MusicPaths.values.length = 0;
     MusicPaths.values.push(...root.musicFiles.map((file) => ({name: file.modPath})));