+++ CHANGELOG.md
@@ -80,3 +80,4 @@ All notable changes to the "ftlXml" extension will be documented in this file.
 - Add support for completing and validating sound file paths
 - add flavorType[id] is text ref
 - Add support for completing, validating, go to for animations
+- Add support for img paths for animations